# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StreamFlix API - Netflix-style streaming & movie discovery platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List catalog items with optional filtering
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [trending, popular, rating, new, releaseYear]
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsPage"
  /items/trending:
    get:
      operationId: getTrending
      tags: [items]
      summary: Get trending items by recent interaction velocity
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: genre
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Trending items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/popular:
    get:
      operationId: getPopular
      tags: [items]
      summary: Get top popular items
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: genre
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Popular items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/new-releases:
    get:
      operationId: getNewReleases
      tags: [items]
      summary: Get recently added or released items
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: New releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
...
  /items/genre/{genre}:
    get:
      operationId: getItemsByGenre
      tags: [items]
      summary: Get movies in a specific genre
      parameters:
        - name: genre
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [popular, rating, new, trending]
            default: popular
      responses:
        "200":
          description: Genre items page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsPage"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get a single item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{id}/trailer:
    get:
      operationId: getItemTrailer
      tags: [items]
      summary: Get trailer information for an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trailer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrailerInfo"
        "404":
          description: Trailer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Admin-only item CRUD
  /admin/items:
    post:
      operationId: createItem
      tags: [admin]
      summary: Create a catalog item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /admin/items/{id}:
    patch:
      operationId: updateItem
      tags: [admin]
      summary: Update a catalog item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [admin]
      summary: Delete a catalog item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── WATCH PROGRESS ─────────────────────────────────────────────────────────
...
  /recommendations/similar/{itemId}:
    get:
      operationId: similarItems
      tags: [recommendations]
      summary: Get items similar to a given item (content-based / ALS hybrid)
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Similar items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
...
  /recommendations/top-picks:
    get:
      operationId: getTopPicks
      tags: [recommendations]
      summary: Get personalized top picks for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Top picks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"

  # ─── INTERACTIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    Item:
      type: object
      required: [id, title, description, genres, tags, rating, releaseYear, createdAt]
      properties:
        id:
          type: integer
        tmdbId:
          type: ["integer", "null"]
        title:
          type: string
        tagline:
          type: ["string", "null"]
        description:
          type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        voteCount:
          type: ["integer", "null"]
        popularity:
          type: ["number", "null"]
        releaseYear:
          type: integer
        releaseDate:
          type: ["string", "null"]
        posterUrl:
          type: ["string", "null"]
        backdropUrl:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        trailerUrl:
          type: ["string", "null"]
        trailerKey:
          type: ["string", "null"]
        trailerSite:
          type: ["string", "null"]
        trailerName:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        director:
          type: ["string", "null"]
        cast:
          type: array
          items:
            type: string
        originalLanguage:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        type:
          type: string
          enum: [movie, series, documentary]
        interactionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ItemInput:
      type: object
      required: [title, description, genres, tags, rating, releaseYear, type]
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        tagline:
          type: string
        description:
          type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        voteCount:
          type: integer
        popularity:
          type: number
        releaseYear:
          type: integer
        releaseDate:
          type: string
        posterUrl:
          type: string
        backdropUrl:
          type: string
        logoUrl:
          type: string
        trailerUrl:
          type: string
        trailerKey:
          type: string
        trailerSite:
          type: string
        trailerName:
          type: string
        duration:
          type: integer
        director:
          type: string
        cast:
          type: array
          items:
            type: string
        originalLanguage:
          type: string
        country:
          type: string
        type:
          type: string
          enum: [movie, series, documentary]
    ItemUpdate:
      type: object
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        tagline:
          type: string
        description:
          type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        voteCount:
          type: integer
        popularity:
          type: number
        releaseYear:
          type: integer
        releaseDate:
          type: string
        posterUrl:
          type: string
        backdropUrl:
          type: string
        logoUrl:
          type: string
        trailerUrl:
          type: string
        trailerKey:
          type: string
        trailerSite:
          type: string
        trailerName:
          type: string
        duration:
          type: integer
        director:
          type: string
        cast:
          type: array
          items:
            type: string
        originalLanguage:
          type: string
        country:
          type: string
        type:
          type: string
          enum: [movie, series, documentary]
    ItemsPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TrailerInfo:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
        key:
          type: ["string", "null"]
        site:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        embedUrl:
          type: ["string", "null"]

    # ── Watch Progress ────────────────────────────────────────────────────────
...